+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added stage checks to presolve, freereoptsolve, freetransform
 - Added primal_dual_evolution recipe and a plot recipe
 ### Fixed
+- Only redirect stdout and stderr in redirectOutput() so that file output still works afterwards
 ### Changed
 - GitHub actions using Mac now use precompiled SCIP from latest release
 ### Removed